refactored server.js  , error handling
