+++ app.js
@@ -1,0 +1,28 @@
+const express = require("express");
+const app = express();
+
+
+const { errorHandler } = require("./middleware/errorHandler");
+const path = require("path");
+require("dotenv").config();
+const cors = require("cors");
+
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
+app.use(cors());
+
+app.use("/api/products", require("./routes/api/products"));
+app.use("/api/users", require("./routes/api/users"));
+app.use("/api/orders", require("./routes/api/orders"));
+app.use("/api/checkout", require("./routes/api/checkout"));
+
+
+
+app.use(express.static(path.join(__dirname, "./frontend/build")));
+app.get("*", (req, res) => {
+  res.sendFile(path.join(__dirname, "./frontend/build/index.html"));
+});
+
+app.use(errorHandler);
+
+module.exports = app;
